+++ changelog.md
@@ -1,3 +1,7 @@
+### 4.5.0.62-experimental
++ fix missing tag (jprik, cflip)
++ add an event for contributor
+
 ### 4.5.0.61-experimental
 + Implement Diabolo Lib
 + Add osmium ore